refactor: web용 패키지 다시 분리
