+++ src/main/java/chess/ConsoleApplication.java
@@ -1,4 +1,6 @@
-package chess.console;
+package chess;
+
+import chess.console.ChessGame;
 
 public class ConsoleApplication {
     public static void main(String[] args) {
