+++ src/main/java/chess/domain/GameSwitch.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain;
 
 public enum GameSwitch {
     ON,
