+++ src/main/java/chess/web/dao/board/BoardDao.java
@@ -1,4 +1,4 @@
-package chess.web.dao;
+package chess.web.dao.board;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
