+++ src/main/java/chess/web/dao/member/MemberDao.java
@@ -1,4 +1,4 @@
-package chess.web.dao;
+package chess.web.dao.member;
 
 import chess.web.Member;
 import java.util.List;
