+++ src/test/java/chess/web/dao/board/BoardDaoTest.java
@@ -1,4 +1,4 @@
-package chess.web.dao;
+package chess.web.dao.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
