+++ src/test/java/chess/web/service/MemberServiceTest.java
@@ -3,7 +3,7 @@ package chess.web.service;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import chess.web.dao.MemberDaoImpl;
+import chess.web.dao.member.MemberDaoImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
